rename: fetch 유틸함수 remote 디렉터리로 이동
