+++ src/views/Payment/recoil/orderItemState.ts
@@ -1,9 +1,10 @@
 import credentialState from '@recoil/server/credentialState';
 import serverUrlState from '@recoil/server/serverUrlState';
-import { atom, atomFamily, selector, selectorFamily, useRecoilValue } from 'recoil';
-import generateFetchOrders from '../utils/fetchOrders';
+import { atomFamily, selectorFamily, useRecoilValue } from 'recoil';
+
 import { ORDER_PATH } from '@constants/urlConstants';
 import { OrderType } from 'types/OrderType';
+import generateFetchOrders from '../remote/fetchOrders';
 
 const orderItemState = atomFamily<OrderType, number>({
   key: 'orderItemState',
